feat(chat): implement infinite scroll
