--- src/components/ChatBoxContent/index.ts
@@ -1,2 +1,0 @@
-export * from './ChatBoxContent';
-export { default } from './ChatBoxContent';
